move rule export tests related logic to the test itself
